+++ demo/HuggingFace/tests/test_interface.py
@@ -44,12 +44,11 @@ def test_network_result(inetwork):
         output_tensor=[],
         semantic_output="hello",
         median_runtime=9001,
-        models=[],
     )
 
 
 def test_network_checkpoint_result(inetwork):
-    inetwork.NetworkCheckpointResult(network_results=[], accuracy=9001.0, perplexity=5.0)
+    inetwork.NetworkCheckpointResult(network_results=[], accuracy=9001.0, perplexity=5.0, models="")
 
 
 def test_precision(inetwork):